# [processor/lsminterval]Fix incorrect overflow encoding