--- a/processor/lsmintervalprocessor/testdata/merger/README.md
+++ b/processor/lsmintervalprocessor/testdata/merger/README.md
@@ -1,0 +1,9 @@
+## Test data for validating metrics merge
+
+The test data in this directory is for testing merging for `Value#MergeMetric` method. The directory structure is:
+
+```txt
+<test_name>
+  | `input.yaml` (the input `pmetric.Metrics`)
+  | `output.yaml` (the final output of the merge operation)
+```