The binary encoding for the trackers was broken as the trackers did not follow the ordered structure of the `pmetric.Metrics` correctly. This PR fixes this by making the trackers follow a tree like datastructure - similar to `pmetric.Metrics`. The PR also adds a separate test for `merger#Value` to catch such bugs - the older test didn't capture these bugs as the tests for overflow didn't have multiple resource metrics.

Benchmarks have went down a bit since the hierarchical representation of trackers causes more allocations than the linear one:

```diff
name                                                         old time/op    new time/op    delta
pkg:github.com/elastic/opentelemetry-collector-components/processor/lsmintervalprocessor goos:darwin goarch:arm64
Aggregation/sum_cumulative-10                                  10.0µs ± 8%    11.4µs ± 4%  +14.71%  (p=0.008 n=5+5)
Aggregation/sum_delta-10                                       10.6µs ± 2%    11.3µs ± 6%     ~     (p=0.056 n=5+5)
Aggregation/histogram_cumulative-10                            10.7µs ± 1%    11.7µs ± 4%   +9.42%  (p=0.008 n=5+5)
Aggregation/histogram_delta-10                                 11.0µs ± 2%    11.6µs ± 3%   +5.28%  (p=0.008 n=5+5)
Aggregation/exphistogram_cumulative-10                         11.0µs ± 6%    12.1µs ± 5%   +9.72%  (p=0.016 n=5+5)
Aggregation/exphistogram_delta-10                              11.4µs ± 6%    12.2µs ± 5%     ~     (p=0.056 n=5+5)
Aggregation/summary_enabled-10                                 10.5µs ± 4%    11.4µs ± 4%   +7.64%  (p=0.008 n=5+5)
Aggregation/summary_passthrough-10                              911ns ±17%     897ns ± 4%     ~     (p=1.000 n=5+5)
AggregationWithOTTL/sum_cumulative-10                          11.2µs ± 4%    11.6µs ± 2%     ~     (p=0.095 n=5+5)
AggregationWithOTTL/sum_delta-10                               11.2µs ± 2%    11.8µs ± 2%   +5.11%  (p=0.008 n=5+5)
AggregationWithOTTL/histogram_cumulative-10                    12.5µs ± 2%    13.5µs ± 4%   +8.13%  (p=0.008 n=5+5)
AggregationWithOTTL/histogram_delta-10                         12.7µs ± 3%    13.4µs ± 4%   +5.23%  (p=0.016 n=5+5)
AggregationWithOTTL/exphistogram_cumulative-10                 12.6µs ±11%    12.6µs ±10%     ~     (p=0.690 n=5+5)
AggregationWithOTTL/exphistogram_delta-10                      13.2µs ±11%    12.5µs ±13%     ~     (p=0.421 n=5+5)
AggregationWithOTTL/summary_enabled-10                         11.6µs ±10%    12.7µs ± 6%     ~     (p=0.095 n=5+5)
AggregationWithOTTL/summary_passthrough-10                      866ns ± 3%     910ns ± 6%     ~     (p=0.190 n=4+5)
pkg:github.com/elastic/opentelemetry-collector-components/processor/lsmintervalprocessor/internal/merger goos:darwin goarch:arm64
Merge/without_overflow/sum_cumulative_overflow-10               856ns ± 1%     859ns ± 1%     ~     (p=0.421 n=5+5)
Merge/without_overflow/sum_delta_overflow-10                    897ns ± 3%     891ns ± 1%     ~     (p=1.000 n=5+5)
Merge/without_overflow/histogram_cumulative_overflow-10         866ns ± 3%     848ns ± 0%   -2.05%  (p=0.016 n=5+5)
Merge/without_overflow/histogram_delta_overflow-10             1.03µs ± 3%    1.01µs ± 0%     ~     (p=0.214 n=5+5)
Merge/without_overflow/exphistogram_cumulative_overflow-10      865ns ± 2%     848ns ± 0%   -1.96%  (p=0.008 n=5+5)
Merge/without_overflow/exphistogram_delta_overflow-10          1.02µs ± 5%    1.02µs ± 3%     ~     (p=0.794 n=5+5)
Merge/with_overflow/sum_cumulative_overflow-10                  445ns ± 1%     445ns ± 1%     ~     (p=0.905 n=4+5)
Merge/with_overflow/sum_delta_overflow-10                       453ns ± 2%     450ns ± 1%     ~     (p=0.548 n=5+5)
Merge/with_overflow/histogram_cumulative_overflow-10            457ns ± 5%     441ns ± 0%   -3.48%  (p=0.008 n=5+5)
Merge/with_overflow/histogram_delta_overflow-10                 494ns ± 3%     488ns ± 0%     ~     (p=0.206 n=5+5)
Merge/with_overflow/exphistogram_cumulative_overflow-10         448ns ± 2%     443ns ± 1%     ~     (p=0.841 n=5+5)
Merge/with_overflow/exphistogram_delta_overflow-10              498ns ± 2%     494ns ± 0%     ~     (p=0.690 n=5+5)
Marshal/sum_cumulative_overflow-10                              471ns ± 3%     461ns ± 0%     ~     (p=0.087 n=5+5)
Marshal/sum_delta_overflow-10                                   484ns ± 1%     461ns ± 0%   -4.56%  (p=0.008 n=5+5)
Marshal/histogram_cumulative_overflow-10                        656ns ± 5%     629ns ± 1%     ~     (p=0.310 n=5+5)
Marshal/histogram_delta_overflow-10                             632ns ± 8%     606ns ± 1%   -4.23%  (p=0.008 n=5+5)
Marshal/exphistogram_cumulative_overflow-10                     820ns ± 4%     809ns ± 1%     ~     (p=0.310 n=5+5)
Marshal/exphistogram_delta_overflow-10                          831ns ± 4%     816ns ± 0%   -1.82%  (p=0.032 n=5+5)
AppendBinaryReuseBuffer/sum_cumulative_overflow-10              374ns ± 3%     370ns ± 1%     ~     (p=0.056 n=5+5)
AppendBinaryReuseBuffer/sum_delta_overflow-10                   374ns ± 1%     368ns ± 0%   -1.43%  (p=0.008 n=5+5)
AppendBinaryReuseBuffer/histogram_cumulative_overflow-10        478ns ± 0%     476ns ± 0%     ~     (p=0.056 n=5+5)
AppendBinaryReuseBuffer/histogram_delta_overflow-10             480ns ± 1%     476ns ± 1%   -0.73%  (p=0.032 n=5+5)
AppendBinaryReuseBuffer/exphistogram_cumulative_overflow-10     704ns ± 3%     692ns ± 0%   -1.64%  (p=0.008 n=5+5)
AppendBinaryReuseBuffer/exphistogram_delta_overflow-10          740ns ± 4%     700ns ± 0%   -5.34%  (p=0.008 n=5+5)

[Geo mean]                                                     1.73µs         1.75µs        +1.07%

name                                                         old alloc/op   new alloc/op   delta
pkg:github.com/elastic/opentelemetry-collector-components/processor/lsmintervalprocessor goos:darwin goarch:arm64
Aggregation/sum_cumulative-10                                  19.1kB ± 3%    20.2kB ± 2%   +5.85%  (p=0.008 n=5+5)
Aggregation/sum_delta-10                                       18.8kB ± 6%    20.2kB ± 4%   +7.06%  (p=0.032 n=5+5)
Aggregation/histogram_cumulative-10                            20.1kB ± 2%    20.9kB ± 1%   +3.92%  (p=0.008 n=5+5)
Aggregation/histogram_delta-10                                 20.0kB ± 2%    20.6kB ± 2%   +3.28%  (p=0.016 n=5+5)
Aggregation/exphistogram_cumulative-10                         21.0kB ± 2%    22.1kB ± 1%   +5.17%  (p=0.008 n=5+5)
Aggregation/exphistogram_delta-10                              21.3kB ± 3%    22.0kB ± 2%   +3.29%  (p=0.008 n=5+5)
Aggregation/summary_enabled-10                                 19.3kB ± 3%    19.9kB ± 3%   +3.34%  (p=0.032 n=5+5)
Aggregation/summary_passthrough-10                             1.36kB ± 0%    1.36kB ± 1%     ~     (p=0.794 n=5+5)
AggregationWithOTTL/sum_cumulative-10                          16.4kB ± 0%    16.7kB ± 1%   +1.54%  (p=0.008 n=5+5)
AggregationWithOTTL/sum_delta-10                               16.4kB ± 0%    16.6kB ± 0%   +1.03%  (p=0.008 n=5+5)
AggregationWithOTTL/histogram_cumulative-10                    23.5kB ± 2%    24.3kB ± 0%   +3.04%  (p=0.008 n=5+5)
AggregationWithOTTL/histogram_delta-10                         23.8kB ± 1%    24.3kB ± 1%   +1.88%  (p=0.008 n=5+5)
AggregationWithOTTL/exphistogram_cumulative-10                 20.6kB ±21%    17.5kB ± 2%     ~     (p=0.111 n=5+4)
AggregationWithOTTL/exphistogram_delta-10                      24.9kB ± 1%    17.7kB ± 2%  -29.04%  (p=0.029 n=4+4)
AggregationWithOTTL/summary_enabled-10                         22.3kB ± 1%    20.7kB ±19%     ~     (p=0.690 n=5+5)
AggregationWithOTTL/summary_passthrough-10                     1.36kB ± 1%    1.37kB ± 1%     ~     (p=0.222 n=5+5)
pkg:github.com/elastic/opentelemetry-collector-components/processor/lsmintervalprocessor/internal/merger goos:darwin goarch:arm64
Merge/without_overflow/sum_cumulative_overflow-10               0.00B          0.00B          ~     (all equal)
Merge/without_overflow/sum_delta_overflow-10                    24.0B ± 0%     24.0B ± 0%     ~     (all equal)
Merge/without_overflow/histogram_cumulative_overflow-10         0.00B          0.00B          ~     (all equal)
Merge/without_overflow/histogram_delta_overflow-10              24.0B ± 0%     24.0B ± 0%     ~     (all equal)
Merge/without_overflow/exphistogram_cumulative_overflow-10      0.00B          0.00B          ~     (all equal)
Merge/without_overflow/exphistogram_delta_overflow-10           24.0B ± 0%     24.0B ± 0%     ~     (all equal)
Merge/with_overflow/sum_cumulative_overflow-10                  0.00B          0.00B          ~     (all equal)
Merge/with_overflow/sum_delta_overflow-10                       8.00B ± 0%     8.00B ± 0%     ~     (all equal)
Merge/with_overflow/histogram_cumulative_overflow-10            0.00B          0.00B          ~     (all equal)
Merge/with_overflow/histogram_delta_overflow-10                 8.00B ± 0%     8.00B ± 0%     ~     (all equal)
Merge/with_overflow/exphistogram_cumulative_overflow-10         0.00B          0.00B          ~     (all equal)
Merge/with_overflow/exphistogram_delta_overflow-10              8.00B ± 0%     8.00B ± 0%     ~     (all equal)
Marshal/sum_cumulative_overflow-10                               712B ± 0%      712B ± 0%     ~     (all equal)
Marshal/sum_delta_overflow-10                                    712B ± 0%      712B ± 0%     ~     (all equal)
Marshal/histogram_cumulative_overflow-10                       1.38kB ± 0%    1.38kB ± 0%     ~     (all equal)
Marshal/histogram_delta_overflow-10                            1.32kB ± 0%    1.32kB ± 0%     ~     (all equal)
Marshal/exphistogram_cumulative_overflow-10                    1.32kB ± 0%    1.32kB ± 0%     ~     (all equal)
Marshal/exphistogram_delta_overflow-10                         1.34kB ± 0%    1.34kB ± 0%     ~     (all equal)
AppendBinaryReuseBuffer/sum_cumulative_overflow-10               320B ± 0%      320B ± 0%     ~     (all equal)
AppendBinaryReuseBuffer/sum_delta_overflow-10                    320B ± 0%      320B ± 0%     ~     (all equal)
AppendBinaryReuseBuffer/histogram_cumulative_overflow-10         640B ± 0%      640B ± 0%     ~     (all equal)
AppendBinaryReuseBuffer/histogram_delta_overflow-10              640B ± 0%      640B ± 0%     ~     (all equal)
AppendBinaryReuseBuffer/exphistogram_cumulative_overflow-10      800B ± 0%      800B ± 0%     ~     (all equal)
AppendBinaryReuseBuffer/exphistogram_delta_overflow-10           816B ± 0%      816B ± 0%     ~     (all equal)

[Geo mean]                                                     1.51kB         1.50kB        -0.55%

name                                                         old allocs/op  new allocs/op  delta
pkg:github.com/elastic/opentelemetry-collector-components/processor/lsmintervalprocessor goos:darwin goarch:arm64
Aggregation/sum_cumulative-10                                     195 ± 2%       223 ± 2%  +14.33%  (p=0.008 n=5+5)
Aggregation/sum_delta-10                                          194 ± 6%       225 ± 3%  +16.07%  (p=0.008 n=5+5)
Aggregation/histogram_cumulative-10                               192 ± 2%       215 ± 1%  +11.99%  (p=0.008 n=5+5)
Aggregation/histogram_delta-10                                    193 ± 1%       215 ± 1%  +11.39%  (p=0.008 n=5+5)
Aggregation/exphistogram_cumulative-10                            211 ± 1%       241 ± 0%  +14.12%  (p=0.008 n=5+5)
Aggregation/exphistogram_delta-10                                 214 ± 2%       239 ± 2%  +11.58%  (p=0.008 n=5+5)
Aggregation/summary_enabled-10                                    198 ± 2%       224 ± 3%  +12.90%  (p=0.008 n=5+5)
Aggregation/summary_passthrough-10                               33.0 ± 0%      33.4 ± 2%     ~     (p=0.444 n=5+5)
AggregationWithOTTL/sum_cumulative-10                             167 ± 0%       181 ± 1%   +8.38%  (p=0.008 n=5+5)
AggregationWithOTTL/sum_delta-10                                  169 ± 0%       182 ± 0%   +7.95%  (p=0.016 n=5+4)
AggregationWithOTTL/histogram_cumulative-10                       228 ± 2%       254 ± 1%  +11.21%  (p=0.008 n=5+5)
AggregationWithOTTL/histogram_delta-10                            232 ± 1%       257 ± 1%  +10.77%  (p=0.008 n=5+5)
AggregationWithOTTL/exphistogram_cumulative-10                    210 ±22%       194 ± 0%     ~     (p=0.714 n=5+4)
AggregationWithOTTL/exphistogram_delta-10                         254 ± 1%       194 ± 0%  -23.67%  (p=0.029 n=4+4)
AggregationWithOTTL/summary_enabled-10                            235 ± 1%       231 ±20%     ~     (p=0.722 n=5+5)
AggregationWithOTTL/summary_passthrough-10                       33.0 ± 0%      33.6 ± 2%     ~     (p=0.167 n=5+5)
pkg:github.com/elastic/opentelemetry-collector-components/processor/lsmintervalprocessor/internal/merger goos:darwin goarch:arm64
Merge/without_overflow/sum_cumulative_overflow-10                0.00           0.00          ~     (all equal)
Merge/without_overflow/sum_delta_overflow-10                     3.00 ± 0%      3.00 ± 0%     ~     (all equal)
Merge/without_overflow/histogram_cumulative_overflow-10          0.00           0.00          ~     (all equal)
Merge/without_overflow/histogram_delta_overflow-10               3.00 ± 0%      3.00 ± 0%     ~     (all equal)
Merge/without_overflow/exphistogram_cumulative_overflow-10       0.00           0.00          ~     (all equal)
Merge/without_overflow/exphistogram_delta_overflow-10            3.00 ± 0%      3.00 ± 0%     ~     (all equal)
Merge/with_overflow/sum_cumulative_overflow-10                   0.00           0.00          ~     (all equal)
Merge/with_overflow/sum_delta_overflow-10                        1.00 ± 0%      1.00 ± 0%     ~     (all equal)
Merge/with_overflow/histogram_cumulative_overflow-10             0.00           0.00          ~     (all equal)
Merge/with_overflow/histogram_delta_overflow-10                  1.00 ± 0%      1.00 ± 0%     ~     (all equal)
Merge/with_overflow/exphistogram_cumulative_overflow-10          0.00           0.00          ~     (all equal)
Merge/with_overflow/exphistogram_delta_overflow-10               1.00 ± 0%      1.00 ± 0%     ~     (all equal)
Marshal/sum_cumulative_overflow-10                               4.00 ± 0%      4.00 ± 0%     ~     (all equal)
Marshal/sum_delta_overflow-10                                    4.00 ± 0%      4.00 ± 0%     ~     (all equal)
Marshal/histogram_cumulative_overflow-10                         4.00 ± 0%      4.00 ± 0%     ~     (all equal)
Marshal/histogram_delta_overflow-10                              4.00 ± 0%      4.00 ± 0%     ~     (all equal)
Marshal/exphistogram_cumulative_overflow-10                      10.0 ± 0%      10.0 ± 0%     ~     (all equal)
Marshal/exphistogram_delta_overflow-10                           10.0 ± 0%      10.0 ± 0%     ~     (all equal)
AppendBinaryReuseBuffer/sum_cumulative_overflow-10               1.00 ± 0%      1.00 ± 0%     ~     (all equal)
AppendBinaryReuseBuffer/sum_delta_overflow-10                    1.00 ± 0%      1.00 ± 0%     ~     (all equal)
AppendBinaryReuseBuffer/histogram_cumulative_overflow-10         1.00 ± 0%      1.00 ± 0%     ~     (all equal)
AppendBinaryReuseBuffer/histogram_delta_overflow-10              1.00 ± 0%      1.00 ± 0%     ~     (all equal)
AppendBinaryReuseBuffer/exphistogram_cumulative_overflow-10      7.00 ± 0%      7.00 ± 0%     ~     (all equal)
AppendBinaryReuseBuffer/exphistogram_delta_overflow-10           7.00 ± 0%      7.00 ± 0%     ~     (all equal)

[Geo mean]                                                       18.3           18.8        +2.66%
```