@lahsivjar 2025-03-18
[**For reviewers**] Currently, the overflow handling is not enabled anywhere but I have created a card to enable overflow handling configuration in the elastic APM connector: https://github.com/elastic/opentelemetry-collector-components/issues/465. Not sure about the priority on this though so please add any comments to the issue if you have a timeline on when it should be enabled.